beta v1, build 0.02

+++ src/components/GalleryGrid.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import Swal from 'sweetalert2';
 import styles from '../app/Gallery.module.css';
 import { Photo, SelectedOptions } from '../types';
 
@@ -22,7 +23,25 @@ const GalleryGrid: React.FC<GalleryGridProps> = ({
 }) => {
   function handleClick(photo: Photo) {
     if (isSelected(photo)) {
-      handleDeleteSelection(photo);
+      Swal.fire({
+        title: 'Вы уверены?',
+        text: "Вы действительно хотите отменить выбор?",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Да, отменить выбор',
+        cancelButtonText: 'Отмена'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          handleDeleteSelection(photo);
+          Swal.fire(
+            'Удалено!',
+            'Ваш выбор был отменен.',
+            'success'
+          )
+        }
+      })
     } else {
       handleSelectPhoto(photo);
     }
@@ -32,23 +51,26 @@ const GalleryGrid: React.FC<GalleryGridProps> = ({
     return !!selectedOptionsMap[photo.id];
   }
 
-  console.log("gallery grid", photos, savedPhotos);
-
-
   return (
     <div className={styles.galleryGrid}>
       {photos.map((photo, index) => {
-        console.log(selectedOptionsMap[photo.id], photo.id, isSelected(photo));
-                const selectedOptions = selectedOptionsMap[photo.id] || {
+        const selectedOptions = selectedOptionsMap[photo.id] || {
           lastName: '',
           photo10x15: 0,
+          photo15x21: 0,
           photo20x30: 0,
           photoInYearbook: false,
           additionalPhotos: false,
           vignette: false,
           photo10x15Name: '',
+          photo15x21Name: '',
           photo20x30Name: '',
           photoInAlbum: false,
+          allPhotosDigital: false,
+          portraitAlbum2: false,
+          portraitAlbum3: false,
+          singlePhotoDigital: false,
+          photoInCube: false,
         };
 
         return (
@@ -68,9 +90,15 @@ const GalleryGrid: React.FC<GalleryGridProps> = ({
                   <div className="title">ЗАКАЗАНО: {selectedOptions.lastName}</div>
                   {selectedOptions.photoInYearbook && <div className="detailItem">✔ Фото в летопись</div>}
                   {selectedOptions.vignette && <div className="detailItem">✔ ВИНЬЕТКА</div>}
-                  {!!selectedOptions.photoInAlbum && <div className="detailItem">✔ Фото в альбом</div>}
+                  {selectedOptions.photoInAlbum && <div className="detailItem">✔ Фото в альбом</div>}
                   {!!selectedOptions.photo10x15 && <div className="detailItem">✔ Доп фото 10x15 {selectedOptions.photo10x15}шт</div>}
+                  {!!selectedOptions.photo15x21 && <div className="detailItem">✔ Доп фото 15x21 {selectedOptions.photo15x21}шт</div>}
                   {!!selectedOptions.photo20x30 && <div className="detailItem">✔ Доп фото 20x30 {selectedOptions.photo20x30}шт</div>}
+                  {selectedOptions.allPhotosDigital && <div className="detailItem">✔ Все фото в электронном виде</div>}
+                  {selectedOptions.portraitAlbum2 && <div className="detailItem">✔ Портрет в альбом 2</div>}
+                  {selectedOptions.portraitAlbum3 && <div className="detailItem">✔ Портрет в альбом 3</div>}
+                  {selectedOptions.singlePhotoDigital && <div className="detailItem">✔ 1 фото в электронном виде</div>}
+                  {selectedOptions.photoInCube && <div className="detailItem">✔ Фото в кубике</div>}
                 </div>
               </div>
             )}
